Create basic etch-a-sketch with black color
I created an html document with one button to reset the board.
With javascript I created a div "container" and with css I set heighth
and width to 600px and also made it a flex-container with
row-direction.
Then I used javascript to put 16 "column" divs into the container, which
have a flex-grow: 1 so they take all the available space on the main
axis of the container evenly. I also set and flex-direction: column.
Then I added 16 "pixel" divs to each container, which also have a
flex-grow: 1, so they take all the space on the "column" main axis
evenly.

Now that the grid was ready, I added an eventListener to each of the
"pixel" divs during their creation so that they change their background
color to black on mouseover.

At last I created a function that removes all the columns from the
container and creates everything new to reset the etch-a-sketch and
linked it to the "Reset" button.