Add rainbow and black modified
I added 2 buttons to my HTML to select the color for each mouseover.
Black mode sets the style.backgroundcolor to black.
Rainbow mode creates 3 variables which are filled with a random number
from 0 to 255.
Then it sets the style.backgroundcolor to rgb(random,random,random), so
each time the mouse passes over the pixel, it gets a new random color.